bottom left socials are good on desktop renamed some of the header elements broke them out individually rm'd unused css
