+++ components/header/subscribe.tsx
@@ -29,20 +29,20 @@ const Subscribe = () => {
   };
 
   return (
-    <div className='headline-menu-right'>
+    <div className='newsletter-subscribe'>
       <form className='email-form' onSubmit={handleSubmit}>
         {/* <label className=''>Newsletter</label> */}
+        {status === 'Subscription successful' && <div className='email-form-success'>Thanks talk soon!</div>}
+        {status === 'Subscription failed' && <div className='email-form-error'>Please try that again!</div>}
         <input 
           type="email" 
           name="email" 
           value={email}
-          placeholder='Newsletter signup'
+          placeholder='You @ Email'
           onChange={e => setEmail(e.target.value)}
         />
         
         <button type="submit" value="Subscribe">Subscribe</button>
-        {status === 'Subscription successful' && <p>Thanks talk soon!</p>}
-        {status === 'Subscription failed' && <p>Please try that again!</p>}
       </form>
     </div>
   );
